requirements.txt
advanced_alchemy==0.13.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
ecdsa==0.19.0
exceptiongroup==1.2.1
Faker==25.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
litestar==2.9.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
polyfactory==2.16.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
PyYAML==6.0.1
rich==13.7.1
rich-click==1.8.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
uvicorn==0.30.1
xxhash==3.4.1
